+++ src/cucumber/utils/parseValue.ts
@@ -1,46 +1,44 @@
-export enum ParseValueType {
-  Number = "number",
-  BigInt = "bigint",
-  Boolean = "boolean",
-  Object = "object",
-  String = "string",
-}
+import { ERROR_MESSAGES } from "../../internal/utils/errorMessages.js";
 
-const getConversionErrorMessage = (type: ParseValueType, value: string) =>
-  `Cannot convert "${value}" to actual value's type: ${type}`;
+export type ParseValueType =
+  | "number"
+  | "bigint"
+  | "boolean"
+  | "object"
+  | "string";
 
 export const parseValue = (
   value: string,
   targetType: ParseValueType,
 ): unknown => {
   switch (targetType) {
-    case ParseValueType.Number:
+    case "number":
       if (!Number.isNaN(+value)) {
         return +value;
       }
-      throw new Error(getConversionErrorMessage(ParseValueType.Number, value));
-    case ParseValueType.BigInt:
+      throw new Error(ERROR_MESSAGES.TYPE_CONVERSION_ERROR(targetType, value));
+    case "bigint":
       try {
         return BigInt(value);
       } catch {
         throw new Error(
-          getConversionErrorMessage(ParseValueType.Object, value),
+          ERROR_MESSAGES.TYPE_CONVERSION_ERROR(targetType, value),
         );
       }
-    case ParseValueType.Boolean:
+    case "boolean":
       if (value === "true" || value === "false") {
         return value === "true";
       }
-      throw new Error(getConversionErrorMessage(ParseValueType.Boolean, value));
-    case ParseValueType.Object:
+      throw new Error(ERROR_MESSAGES.TYPE_CONVERSION_ERROR(targetType, value));
+    case "object":
       try {
         return JSON.parse(value);
       } catch {
         throw new Error(
-          getConversionErrorMessage(ParseValueType.Object, value),
+          ERROR_MESSAGES.TYPE_CONVERSION_ERROR(targetType, value),
         );
       }
-    case ParseValueType.String:
+    case "string":
       return value;
     default:
       throw new Error(`Unsupported type of actual value: ${targetType}`);